refactor: 패키지 이동 및 에러 메시지 수정
request패키지로 관련 테스트를 이동시키고 에러메시지 오탈자를 수정
