+++ src/main/java/vendingmachine/StringConstants.java
@@ -2,7 +2,7 @@ package vendingmachine;
 
 public class StringConstants {
     public static final String PREFIX_OF_ERROR_MESSAGE = "[ERROR] ";
-    public static final String ERROR_MESSAGE_ABOUT_WRONG_CURRENT_BALANCE_INPUT = "보유 금액은 10원 이상어야 하며 10원으로 나누어 떨어지는 정수만 가능합니다";
+    public static final String ERROR_MESSAGE_ABOUT_WRONG_CURRENT_BALANCE_INPUT = "보유 금액은 10원 이상이어야 하며 10원으로 나누어 떨어지는 정수만 가능합니다";
     public static final String ERROR_MESSAGE_ABOUT_WRONG_ITEMS_INVENTORY_INPUT = "하나 이상의 상품에 대한 정보를 입력해 주세요";
     public static final String ERROR_MESSAGE_ABOUT_DUPLICATED_ITEM_NAMES_IN = "상품명은 중복될 수 없습니다";
     public static final String ERROR_MESSAGE_ABOUT_WRONG_ITEM_INVENTORY_INPUT_FORMAT = "아이템 정보는 대괄호('[]')로 묶여 있어야 합니다";
